requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
python-dotenv==1.2.2
supabase==2.28.3
gunicorn==23.0.0